adding a check before the user tries to withdraw their crypto to make sure they can cover their active wagers

+++ BetIt/Session/Models/WagerStatus.swift
@@ -10,7 +10,7 @@ import Foundation
 struct WagerStatus: Decodable, Hashable {
     var wagerId: Int
     var isActive: Bool
-    var amount: Int
+    var amount: Decimal
     var gameStartTime: Date
     var chosenTeam: Int
 }
